resume-formatter-ai/backend-ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-docx
pypdf
docx2txt
requests
